Use map method; move common methods to base Model class
